fix: Added file to hashin hash-object

+++ main.py
@@ -18,21 +18,25 @@ def handle_init(args) :
 
 def handle_hash_object(args) : 
     try : 
-        raw_data = args.string
-        header_string = f"blob {len(raw_data.encode())}\0"
-        blob_bytes = header_string.encode() + raw_data.encode()
-        sha1_hash = hashlib.sha1(blob_bytes).hexdigest()
-        compressed_blob_bytes = zlib.compress(blob_bytes)
-        print(sha1_hash)
+        file_path = pathlib.Path(args.path)
+        if file_path.exists() : 
+            with open(file_path, "r") as f :
+                raw_data = f.read()
+                
+                header_string = f"blob {len(raw_data.encode())}\0"
+                blob_bytes = header_string.encode() + raw_data.encode()
+                sha1_hash = hashlib.sha1(blob_bytes).hexdigest()
+                compressed_blob_bytes = zlib.compress(blob_bytes)
+                print(sha1_hash)
 
-        dir_path = pathlib.Path(f".mgit/objects/{sha1_hash[:2]}")
-        if dir_path.exists() :
-            with open(f".mgit/objects/{sha1_hash[:2]}/{sha1_hash[2:]}", "wb") as f: 
-                f.write(compressed_blob_bytes)
-        else : 
-            os.mkdir(dir_path)
-            with open(f".mgit/objects/{sha1_hash[:2]}/{sha1_hash[2:]}", "wb") as f: 
-                f.write(compressed_blob_bytes)
+                dir_path = pathlib.Path(f".mgit/objects/{sha1_hash[:2]}")
+                if dir_path.exists() :
+                    with open(f".mgit/objects/{sha1_hash[:2]}/{sha1_hash[2:]}", "wb") as f: 
+                        f.write(compressed_blob_bytes)
+                else : 
+                    os.mkdir(dir_path)
+                    with open(f".mgit/objects/{sha1_hash[:2]}/{sha1_hash[2:]}", "wb") as f: 
+                        f.write(compressed_blob_bytes)
     except Exception as e : 
         print(e)  
 
@@ -59,7 +63,7 @@ init_parser.set_defaults(func=handle_init)
 
 # Create hash-object Parser
 hash_object_parser = subparsers.add_parser("hash-object", help="Hash a object")
-hash_object_parser.add_argument("string",type=str)
+hash_object_parser.add_argument("path",type=str)
 hash_object_parser.set_defaults(func=handle_hash_object)
 
 # Create cat-file Parser
